package: fakechroot: add configuration for the environment alpine-make-rootfs
This adds the configuration for the environment alpine-make-rootfs and
its replacement executable for busybox, bbsuid and busybox-extra.

The alpine-make-rootfs runs the busybox's post-install package script
that installs the busybox symlinks to the chroot'ed directory.

	$ /bin/busybox --install -s

There are two things to workaround because of busybox and fakechroot.

First, fakechroot execve()s the hashbang of post-install scripts in the
chroot'ed environment (i.e. #!/bin/sh) and the shell interpreter is an
absolute symlink to /bin/busybox. The symlink is broken if /bin/busybox
is not installed on the host.

For now, there is nothing yet to tell fakechroot to resolve absolute
path from the chrooted environment that point to the host (the opposite
of FAKECHROOT_EXCLUDE_PATH). As a workaround, the environment excludes
/bin/sh from been chrooted and runs the one from the host instead.

Second, the busybox's post-install package script creates all the
symlinks for the applets that are compiled in. Then, these applets are
used by either alpine-make-rootfs or apk or package's scripts and fail
as these applets are absolute symlinks to /bin/busybox which are broken
if /bin/busybox is not installed on the host.

Therefore, the environment substitutes the executable /bin/busybox
(/bin/bbsuid and /bin/busybox-extra) by the replacement for the
executable busybox and assures that applets are accessible by
intercepting busybox invokations with both arguments --install -s and
makes copies of the busybox binary instead creating symlinks.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>