hackfile.mk: copy overlay and excute post-build script
Buildroot performs some modifications to the target directory through
the target target-finalize. That target is skipped since the commit
df700ab (Do not run target-finalize). However, some of the modifications
are required for some configuration such as the copy of the overlay or
the excute of the post build script.

This backports the these two modifications to hackfile.mk.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>