fakechroot: fix expansion of absolute symlinks
The fakechroot command expands all absolute symlinks with the chroot
base directory.

But fakechroot is also used to create a root file-system and it is
required to keep these absolute symlinks not expanded; the symlinks will
be broken once the rootfs is mounted.

Some utilities make absolute symlinks that should not be expanded as
they are local to the rootfs.

For example, systemctl creates absolute symlinks to enable new services
below the directory /etc/systemd. The machine-id is also a symlink.

This introduces the new variable FAKECHROOT_EXCLUDE_SYMLINK that lists
symlinks and directories that are excluded from being expanded with the
chroot base directory. The environments use that variable to disable
expansion for the service symlinks located in /etc/systemd and the
/var/lib/dbus/machine-id file.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>