board: alpinelinux: configure network at bootup
The network is down at bootup that makes Alpine Linux accessible only
physically via a getty spawned either on virtual consoles or on a serial
console.

This enables the networking initscript at bootup and backports the
buildroot's facility to configure a network interface through DHCP which
is setup by the alpinelinux's post-build script instead of dedicated
packages. It appends the interface set by the BR2_SYSTEM_DHCP config to
the network interfaces.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>