fakechroot: add the replacement executable for ldconfig
According to the man page ldconfig(8):

	-r root

	Change to and use root as the root directory.

This adds the replacement for the executable ldconfig that calls the
program from the chroot'ed environment with the option -r root where
root is the path to the chroot directory (aka $FAKECHROOT_BASE_ORIG").

Fixes:

	ldconfig: Can't create temporary cache file /etc/ld.so.cache~: Permission denied
	error: command failed to execute correctly

Note: Additionally, this resets the variable FAKECHROOT_CMD_ORIG in the
replacement for unshare to let its children substitute commands; this is
required to run the replacement for ldconfig. Also, as a workaround,
this configures pacman to run /usr/bin/ldconfig instead of the default
/sbin/ldconfig in order to use $(which ldconfig) in the configuration
for the environment of pacstrap. In the world of Arch Linux, binaries
are installed to /usr/bin only; the file /sbin is a symlink to /usr/sbin
which is a symlink to /usr/bin (and /bin is a symlink to /usr/bin).

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>