board: calamares: resize rootfs to 100%
This adds the calamares module fsresizer to the settings in order to
resize the rootfs partition and filesystem to 100%.

Note: This makes the init-script calamares depend on dbus as kpmcore
requires this daemon up to run jobs. Also, this substitutes the command
dbus-uuidgen by true to prevent the script post-install[1] in the dbus
package to fail; Alpine Linux has not yet support for fakechroot.

[1]: https://git.alpinelinux.org/aports/tree/main/dbus/dbus.post-install

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>