board: plymouth: create users and groups manually
The users and the groups have to be created manually since Alpine Linux
does not have support yet for fakechroot.

This creates the user table[1] that adds manually the users and groups
messagebus[2] and polkitd[3]; they are supposed to be added by the
pre-install script of their package.

[1]: https://buildroot.org/downloads/manual/manual.html#customize-users
[2]: https://git.alpinelinux.org/aports/tree/community/polkit/polkit.pre-install
[3]: https://git.alpinelinux.org/aports/tree/main/dbus/dbus.pre-install

See also c627129 (board: calamares: create users and groups manually).

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>