skeleton-{alpine,arch}linux: install /bin, /lib, and /sbin
These skeletons do not install the required /bin, /lib and /sbin
directories (or symlinks), which may result in an imcomplete tree in
staging and lead to the error below.

This creates the call SYSTEM_BIN_SBIN_LIB_DIRS in hackfile.mk to install
the directories /bin, /sbin and /lib. It reuses the upstream calls to
SYSTEM_USR_SYMLINKS_OR_DIRS & SYSTEM_LIB_SYMLINK to INSTALL_STAGING_CMDS
so the required directories or symlinks are created.

Fixes:

	>>> host-gcc-final 10.3.0 Building
	(...)
	/home/gportay/src/linux-distros-br2/buildroot/output/host/x86_64-buildroot-linux-gnu/bin/ld: cannot find crti.o: No such file or directory
	collect2: error: ld returned 1 exit status

[1]: https://github.com/buildroot/buildroot/commit/efc974e6860484753ae6f1a86f1962d2300c3c16
[2]: https://github.com/buildroot/buildroot/commit/45fbadb0b785de116b369fb1d413063126316c19